requirements.txt
numpy
minisom
scipy
pandas
matplotlib
pillow
torch
torchvision
numba
seaborn
PyYAML
wandb
scikit-learn
tqdm
bokeh
plotly
